+++ ecl_time/src/test/frequency.cpp
@@ -9,6 +9,7 @@
 *****************************************************************************/
 
 #include <gtest/gtest.h>
+#include <iostream>
 #include "../../include/ecl/time/frequency.hpp"
 #include "../../include/ecl/time/sleep.hpp"
 #include "../../include/ecl/time/timestamp.hpp"
@@ -25,10 +26,11 @@
 
 std::ostream& operator <<( std::ostream& ostream , const ecl::FrequencyDiagnostics& diagnostics )
 {
+  ostream <<  std::fixed;
   ostream << "hz : " << diagnostics.hz;
   ostream << ", min : " << diagnostics.minimum_interval;
   ostream << ", max : " << diagnostics.maximum_interval;
-  ostream << ", last_incoming : " << std::fixed << diagnostics.last_incoming << std::scientific;
+  ostream << ", last_incoming : " << diagnostics.last_incoming << std::scientific;
   // could print the others, but that would require use of std ios flags to pretty print correctly.
   // let the user do that how they wish if they want to
   return ostream;
@@ -46,6 +48,9 @@ TEST(FrequencyMonitorTests,silent) {
   ecl::MilliSleep sleep_ms;
   unsigned int period = 100;
 
+//  std::cout << "\n**********************************************************" << std::endl;
+//  std::cout << "* Frequency Monitor Test" << std::endl;
+//  std::cout << "**********************************************************\n" << std::endl;
   float epsilon = 0.3;
   for ( unsigned int i = 0; i < 50; ++i ) {
     monitor.update();
@@ -59,6 +64,7 @@ TEST(FrequencyMonitorTests,silent) {
   }
   std::cout << diagnostics << std::endl;
   EXPECT_NEAR(diagnostics.hz, 5.0, epsilon);
+  SUCCEED();
 }
 
 #endif /* ECL_HAS_TIMESTAMP */
@@ -69,6 +75,6 @@ TEST(FrequencyMonitorTests,silent) {
 
 int main(int argc, char **argv) {
 
-    testing::InitGoogleTest(&argc,argv);
-    return RUN_ALL_TESTS();
+  testing::InitGoogleTest(&argc,argv);
+  return RUN_ALL_TESTS();
 }